+++ editable.js
@@ -4859,7 +4859,14 @@ var Cursor = (function() {
         this.range.insertNode(element);
       },
 
+      /**
+       * Alias for #setVisibleSelection()
+       */
       setSelection: function() {
+        this.setVisibleSelection();
+      },
+
+      setVisibleSelection: function() {
         rangy.getSelection().setSingleRange(this.range);
       },
 
@@ -4885,8 +4892,11 @@ var Cursor = (function() {
        * Get the BoundingClientRect of the cursor.
        * The returned values are absolute to document.body.
        */
-      getCoordinates: function() {
+      getCoordinates: function(positioning) {
+        positioning = positioning || 'absolute';
+
         var coords = this.range.nativeRange.getBoundingClientRect();
+        if (positioning === 'fixed') return coords;
 
         // code from mdn: https://developer.mozilla.org/en-US/docs/Web/API/window.scrollX
         var x = (window.pageXOffset !== undefined) ? window.pageXOffset : (document.documentElement || document.body.parentNode || document.body).scrollLeft;
@@ -4921,6 +4931,9 @@ var Cursor = (function() {
         if (this.isSelection) return new Cursor(this.host, this.range);
       },
 
+      /**
+       * Move the cursor to the beginning of the host.
+       */
       moveAtBeginning: function(element) {
         if (!element) element = this.host;
         this.setHost(element);
@@ -4929,6 +4942,9 @@ var Cursor = (function() {
         if (this.isSelection) return new Cursor(this.host, this.range);
       },
 
+      /**
+       * Move the cursor to the end of the host.
+       */
       moveAtEnd: function(element) {
         if (!element) element = this.host;
         this.setHost(element);
@@ -4937,6 +4953,9 @@ var Cursor = (function() {
         if (this.isSelection) return new Cursor(this.host, this.range);
       },
 
+      /**
+       * Move the cursor after the last visible character of the host.
+       */
       moveAtTextEnd: function(element) {
         return this.moveAtEnd(parser.latestChild(element));
       },
@@ -6159,6 +6178,28 @@ var Selection = (function() {
     expandTo: function(elem) {
       this.range = content.expandTo(this.host, this.range, elem);
       this.setSelection();
+    },
+
+    /**
+     *  Collapse the selection at the beginning of the selection
+     *
+     *  @return Cursor instance
+     */
+    collapseAtBeginning: function(elem) {
+      this.range.collapse(true);
+      this.setSelection();
+      return new Cursor(this.host, this.range);
+    },
+
+    /**
+     *  Collapse the selection at the end of the selection
+     *
+     *  @return Cursor instance
+     */
+    collapseAtEnd: function(elem) {
+      this.range.collapse(false);
+      this.setSelection();
+      return new Cursor(this.host, this.range);
     },
 
     /**