+++ Changelog.md
@@ -1,4 +1,13 @@
 
-# v0.1.0 (2014-01-24)
+# v0.1.2 (2014-03-07)
+
+- [Add selection methods](https://github.com/upfrontIO/Editable.JS/pull/64)
+  - New Selection methods:  
+    #collapseAtBeginning()  
+    #collapseAtEnd()
+  - New Cursor and Selection method:  
+    #setVisibleSelection() (alias for #setSelection())
+
+# v0.1.1 (2014-01-24)
 
 - Setup Versioning